Move PAT scenarios into /scenarios directory
