+++ src/services/production-approval-tests/scenarios/c02-reason-for-no-carrier-registration-number.test.js
@@ -1,7 +1,7 @@
 import { runScenarioC02Tests } from './c02-reason-for-no-carrier-registration-number.js'
-import { buildWasteInput } from './test-helpers.js'
-import { PAT_STATUS } from './status.js'
-import { REASONS_FOR_NO_REGISTRATION_NUMBER } from '../../common/constants/reasons-for-no-registration-number.js'
+import { buildWasteInput } from '../test-helpers.js'
+import { PAT_STATUS } from '../status.js'
+import { REASONS_FOR_NO_REGISTRATION_NUMBER } from '../../../common/constants/reasons-for-no-registration-number.js'
 
 describe('runScenarioC02Tests', () => {
   it('passes when carrier reasonForNoRegistrationNumber is given', () => {
