+++ src/services/production-approval-tests/scenarios/r02-multiple-waste-items.js
@@ -1,4 +1,4 @@
-import { fail, pass } from './status.js'
+import { fail, pass } from '../status.js'
 
 export function runScenarioR02Tests(wasteInput) {
   const wasteItems = wasteInput?.receipt?.movement?.wasteItems ?? []
