+++ src/services/production-approval-tests/scenarios/r03-road-transport.js
@@ -1,4 +1,4 @@
-import { fail, pass } from './status.js'
+import { fail, pass } from '../status.js'
 
 const ROAD = 'Road'
